+++ src/GUI_download_files.py
@@ -21,71 +21,88 @@ from GUI_forms import FileButton, ProceedButton, ChoiceSection
 class ExampleFileDialog(QDialog):
     """a dialog to download example files
     """
-    def __init__(self, settings, log, parent = None):
+
+    def __init__(self, settings, log, parent=None):
         super().__init__(parent)
         self.settings = settings
         self.log = log
         self.setWindowTitle("Download example files")
         self.setWindowIcon(QIcon(general.favicon))
-        self.resize(200,50)
+        self.resize(200, 50)
         self.init_UI()
         self.setModal(True)
         self.show()
-        
+
     def init_UI(self):
         """establish and fill the UI
         """
         layout = QFormLayout(self)
         self.setLayout(layout)
         self.btn_dic = {}
-        
+
         # sequence files:
         seq_lbl = QLabel("Example sequence files:")
         seq_lbl.setStyleSheet(general.label_style_2nd)
         layout.addRow(seq_lbl)
-        
+
         seq_HLA_btn = QPushButton("Download!", self)
         self.btn_dic[seq_HLA_btn] = ("HLA-A", "HLA-A_01-01-01-01.fa")
         seq_HLA_btn.clicked.connect(self.download_file)
         HLA_msg = "Click here to download an example fasta file for an existing HLA-A allele. It can be used as Input for New Sequence."
         seq_HLA_btn.setWhatsThis(HLA_msg)
         layout.addRow(QLabel("HLA-A example:"), seq_HLA_btn)
-        
+
         seq_KIR1_btn = QPushButton("Download!", self)
         self.btn_dic[seq_KIR1_btn] = ("KIR2DL1", "KIR2DL1_0020101.fa")
         seq_KIR1_btn.clicked.connect(self.download_file)
-        seq_KIR1_btn.setWhatsThis("This is an example fasta file for an existing KIR2DL1 allele. It can be used as Input for New Sequence.")
+        seq_KIR1_btn.setWhatsThis(
+            "This is an example fasta file for an existing KIR2DL1 allele. It can be used as Input for New Sequence.")
         layout.addRow(QLabel("KIR example 1 (KIR2DL1 contains a pseudoexon):"), seq_KIR1_btn)
-        
+
         seq_KIR2_btn = QPushButton("Download!", self)
         self.btn_dic[seq_KIR2_btn] = ("KIR2DL4", "KIR2DL4_0010201.fa")
         seq_KIR2_btn.clicked.connect(self.download_file)
-        seq_KIR2_btn.setWhatsThis("This is an example fasta file for an existing KIR2DL4 allele. It can be used as Input for New Sequence.")
+        seq_KIR2_btn.setWhatsThis(
+            "This is an example fasta file for an existing KIR2DL4 allele. It can be used as Input for New Sequence.")
         layout.addRow(QLabel("KIR example 2 (KIR2DL4 contains a deleted exon):"), seq_KIR2_btn)
-        
+
         seq_MICA_btn = QPushButton("Download!", self)
         self.btn_dic[seq_MICA_btn] = ("MICA", "MICA_002_01.fa")
         seq_MICA_btn.clicked.connect(self.download_file)
-        seq_MICA_btn.setWhatsThis("This is an example fasta file for an existing MICA allele. It can be used as Input for New Sequence.")
+        seq_MICA_btn.setWhatsThis(
+            "This is an example fasta file for an existing MICA allele. It can be used as Input for New Sequence.")
         layout.addRow(QLabel("MICA example:"), seq_MICA_btn)
-        
+
         # IPD submission input files:
         ipd_lbl = QLabel("Example input files for IPD submission:")
         ipd_lbl.setStyleSheet(general.label_style_2nd)
         layout.addRow(ipd_lbl)
-        
+
         ENA_reply_btn = QPushButton("Download!", self)
         self.btn_dic[ENA_reply_btn] = ("ENA reply", "fake_ENA_reply.txt")
         ENA_reply_btn.clicked.connect(self.download_file)
-        ENA_reply_btn.setWhatsThis("This file is a truncated version of the file sent by ENA after ID assignment. It can be used as input for IPD file creation of the example sequences.")
+        ENA_reply_btn.setWhatsThis(
+            "This file is a truncated version of the file sent by ENA after ID assignment. It can be used as input for IPD file creation of the example sequences.")
         layout.addRow(QLabel("ENA reply file:"), ENA_reply_btn)
-        
+
         pretypings_btn = QPushButton("Download!", self)
         self.btn_dic[pretypings_btn] = ("pretypings", "pretypings_example.csv")
         pretypings_btn.clicked.connect(self.download_file)
-        pretypings_btn.setWhatsThis("This file contains a list of previously identified alleles for all loci for each sample to be submitted to IPD. It can be used as input for IPD file creation of the example sequences.")
+        pretypings_btn.setWhatsThis(
+            "This file contains a list of previously identified alleles for all loci for each sample to be submitted to IPD. It can be used as input for IPD file creation of the example sequences.")
         layout.addRow(QLabel("Pretypings file:"), pretypings_btn)
-        
+
+        layout.addRow(QLabel(""))
+
+        all_btn = QPushButton("Download!", self)
+        self.btn_dic[all_btn] = ("all (zipped)", "sample_files.zip")
+        all_btn.clicked.connect(self.download_file)
+        all_btn.setWhatsThis(
+            "This is a zip file with all listed sample files.")
+        lbl = QLabel("All sample files (zipped):")
+        lbl.setStyleSheet(general.label_style_2nd)
+        layout.addRow(lbl, all_btn)
+
     @pyqtSlot()
     def download_file(self):
         """downloads the example file corresponding to the sending button from TypeLoader
@@ -95,9 +112,11 @@ class ExampleFileDialog(QDialog):
         myfile = os.path.join("sample_files", filename)
         if not os.path.exists(myfile):
             myfile = os.path.join(self.settings["root_path"], "_general", "sample_files", filename)
+
         if os.path.exists(myfile):
             suggested_path = os.path.join(self.settings["default_saving_dir"], myfile)
-            chosen_path = QFileDialog.getSaveFileName(self, "Download example {} file...".format(designator), suggested_path)[0]
+            chosen_path = \
+                QFileDialog.getSaveFileName(self, "Download example {} file...".format(designator), suggested_path)[0]
             if chosen_path:
                 try:
                     copyfile(myfile, chosen_path)
@@ -105,10 +124,12 @@ class ExampleFileDialog(QDialog):
                 except Exception as E:
                     self.log.error("\t=> Download failed!")
                     self.log.exception(E)
-                    QMessageBox.warning(self, "Dowload failed", "Sorry, I could not download the file you requested:\n\n{}".format(repr(E)))
+                    QMessageBox.warning(self, "Dowload failed",
+                                        "Sorry, I could not download the file you requested:\n\n{}".format(repr(E)))
         else:
             self.log.error("File not found: {}".format(myfile))
-            QMessageBox.warning(self, "File not found", "Sorry, I could not find the file you requested: \n{}".format(myfile))
+            QMessageBox.warning(self, "File not found",
+                                "Sorry, I could not find the file you requested: \n{}".format(myfile))
             self.log.info("\tDownload aborted. :(")
 
 
@@ -138,7 +159,7 @@ class LogFileDialog(QDialog):
         layout = QVBoxLayout(self)
         self.setLayout(layout)
 
-        msg = "If you wish to show the TypeLoader developers details about a current problems, "
+        msg = "If you wish to show the TypeLoader developers details about a current problem, "
         msg += "you can download a log file here.\n"
         layout.addWidget(QLabel(msg))
 
@@ -216,7 +237,7 @@ class LogFileDialog(QDialog):
                 QMessageBox.information(self, "Check for confidential info before sending!", self.warning)
 
 
-#===========================================================
+# ===========================================================
 # functions:
 
 def log_uncaught_exceptions(cls, exception, tb):
@@ -226,13 +247,16 @@ def log_uncaught_exceptions(cls, exception, tb):
     import traceback
     from PyQt5.QtCore import QCoreApplication
     log.critical('{0}: {1}'.format(cls, exception))
-    log.exception(msg = "Uncaught Exception", exc_info = (cls, exception, tb))
-    #TODO: (future) maybe find a way to display the traceback only once, both in console and logfile?
+    log.exception(msg="Uncaught Exception", exc_info=(cls, exception, tb))
+    # TODO: (future) maybe find a way to display the traceback only once, both in console and logfile?
     sys.__excepthook__(cls, exception, traceback)
-    QCoreApplication.exit(1) 
-    
-pass    
-#===========================================================
+    QCoreApplication.exit(1)
+
+
+pass
+
+
+# ===========================================================
 # main:
 
 
@@ -243,11 +267,11 @@ def main():
     settings_dic = GUI_login.get_settings("admin", log)
     app = QApplication(sys.argv)
     sys.excepthook = log_uncaught_exceptions
-    
-    ex = LogFileDialog(settings_dic, log)
+
+    ex = ExampleFileDialog(settings_dic, log)
     ex.show()
     result = app.exec_()
-    
+
     log.info("<End>")
     sys.exit(result)